# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order
      description: Creates a customer order using the selected products and delivery details.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/lookup:
    get:
      operationId: lookupOrder
      tags: [orders]
      summary: Look up an order
      description: Looks up a customer's order using its order number and phone number.
      parameters:
        - name: orderNumber
          in: query
          required: true
          schema:
            type: string
        - name: phone
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    OrderItemInput:
      type: object
      properties:
        variantId:
          type: integer
          minimum: 1
        productSlug:
          type: string
        size:
          type: string
        quantity:
          type: integer
          minimum: 1
      required:
        - variantId
        - productSlug
        - size
        - quantity
    OrderInput:
      type: object
      properties:
        customerName:
          type: string
          minLength: 2
        phone:
          type: string
          minLength: 8
        city:
          type: string
        address:
          type: string
        paymentMethod:
          type: string
          enum: [cod, prepaid]
        walletCreditToUse:
          type: number
          minimum: 0
          maximum: 500
        loyaltyVerificationToken:
          type: string
          maxLength: 1000
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/OrderItemInput"
      required:
        - customerName
        - phone
        - city
        - address
        - paymentMethod
        - items
    OrderItem:
      type: object
      properties:
        productName:
          type: string
        productSlug:
          type: string
        variantId:
          type: integer
        sku:
          type: string
        colorName:
          type: string
        colorHex:
          type: string
        size:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
      required:
        - productName
        - productSlug
        - variantId
        - sku
        - colorName
        - colorHex
        - size
        - quantity
        - unitPrice
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        customerName:
          type: string
        phone:
          type: string
        city:
          type: string
        address:
          type: string
        paymentMethod:
          type: string
        status:
          type: string
          enum: [new, confirmed, processing, packed, ready_to_ship, shipped, delivered, canceled, returned]
        courier:
          type: ["string", "null"]
        zone:
          type: ["string", "null"]
        total:
          type: number
        subtotal:
          type: number
        bundleDiscount:
          type: number
        walletCreditUsed:
          type: number
        loyaltyPointsEarned:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
        updatedAt:
          type: string
        inventoryDeductedAt:
          type: ["string", "null"]
        packedAt:
          type: ["string", "null"]
        shippedAt:
          type: ["string", "null"]
        deliveredAt:
          type: ["string", "null"]
      required:
        - id
        - orderNumber
        - customerName
        - phone
        - city
        - address
        - paymentMethod
        - status
        - courier
        - zone
        - total
        - subtotal
        - bundleDiscount
        - walletCreditUsed
        - loyaltyPointsEarned
        - items
        - createdAt
        - updatedAt
        - inventoryDeductedAt
        - packedAt
        - shippedAt
        - deliveredAt
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error